+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 - [x] Only run transpiler when needed
 - [x] New extension icon
 - [x] Button to create *tsconfig.json*
-- [ ] Honor `"compileOnSave": false`
-- [ ] Make transpiling async
+- [x] Honor `"compileOnSave": false`
+- [x] Make transpiling async
 - [ ] Updated [readme.md](README.md) with screenshots
 
 Features that have a checkmark are complete and available for